Dieter/Dieter/services/iot-ingestor/requirements.txt
paho-mqtt==1.6.1
psycopg[binary]==3.2.1
